+++ SoundWizardTests/AudioCalculatorTests.swift
@@ -12,7 +12,7 @@ class AudioCalculatorTests: XCTestCase {
 
     func testRandomFreq() throws {
         for _ in 0...1000 {
-            let range: ClosedRange<Float> = 40.0...16_000.0
+            let range: ClosedRange<Double> = 40.0...16_000.0
             let randomFreq = AudioMath.randomFreq(in: range, repelEdges: false)
             if !range.contains(randomFreq) {
                 fatalError()
@@ -23,14 +23,14 @@ class AudioCalculatorTests: XCTestCase {
     }
     
     func testDBToPerc() {
-        let dB: Float = 3
+        let dB: Double = 3
         let gain = Gain(dB: dB)
         let percentage = gain.percentage.rounded(places: 1)
         XCTAssertEqual(percentage, 2)
     }
     
     func testPercToDB() {
-        let percentage: Float = 2
+        let percentage: Double = 2
         let dB = Gain(percentage: percentage).dB.rounded(places: 1)
         XCTAssertEqual(dB, 3)
     }
